requirements.txt
asgiref==3.2.3
bootstrap4==0.1.0
Django==3.0
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
factory-boy==2.12.0
Faker==3.0.0
loguru==0.4.0
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
slugify==0.0.1
sqlparse==0.3.0
text-unidecode==1.3
transliterate==1.10.2
